Move backend CAPTCHA code to modules.core
