+++ indico/modules/core/controllers.py
@@ -5,7 +5,13 @@
 # modify it under the terms of the MIT License; see the
 # LICENSE file for more details.
 
+import base64
+import random
+import string
+
 import requests
+from captcha.audio import AudioCaptcha
+from captcha.image import ImageCaptcha
 from flask import current_app, flash, jsonify, redirect, request, session
 from packaging.version import Version
 from pkg_resources import DistributionNotFound, get_distribution
@@ -26,6 +32,7 @@ from indico.core.sentry import submit_user_feedback
 from indico.core.settings import PrefixSettingsProxy
 from indico.modules.admin import RHAdminBase
 from indico.modules.cephalopod import cephalopod_settings
+from indico.modules.core.captcha import get_captcha_plugin
 from indico.modules.core.forms import SettingsForm
 from indico.modules.core.settings import core_settings, social_settings
 from indico.modules.core.views import WPContact, WPSettings
@@ -325,3 +332,29 @@ class RHConfig(RH):
                        has_privacy_policy=bool(privacy_policy_url or privacy_policy_html),
                        privacy_policy_html=privacy_policy_html,
                        privacy_policy_url=privacy_policy_url)
+
+
+class RHAPIGenerateCaptcha(RH):
+    """Generate a CAPTCHA.
+
+    If a CAPTCHA plugin is available it is used,
+    otherwise we fallback to the built-in CAPTCHA.
+
+    See `CaptchaPluginMixin` for more information about
+    custom CAPTCHA plugins.
+    """
+
+    def _generate_captcha(self):
+        """Built-in 4-digit CAPTCHA which supports both image and audio"""
+        answer = ''.join(random.choices(string.digits, k=4))
+        image = ImageCaptcha().generate(answer).read()
+        audio = AudioCaptcha().generate(answer)
+        return answer, {'image': base64.b64encode(image), 'audio': base64.b64encode(audio)}
+
+    def _process_GET(self):
+        if plugin := get_captcha_plugin():
+            answer, data = plugin.generate_captcha()
+        else:
+            answer, data = self._generate_captcha()
+        session['captcha_answer'] = answer
+        return jsonify(data)
