+++ indico/modules/events/registration/blueprint.py
@@ -212,9 +212,6 @@ _bp_participation.add_url_rule('/manage/participants/', 'manage', regforms.RHMan
 _bp.add_url_rule('/manage/registration/<int:reg_form_id>/privacy/settings', 'manage_registration_privacy_settings',
                  privacy.RHRegistrationPrivacy, methods=('GET', 'POST'))
 
-# CAPTCHA
-_bp.add_url_rule('!/api/captcha/generate', 'api_generate_captcha', api.RHAPIGenerateCaptcha)
-
 # Legacy URLs
 _compat_bp = IndicoBlueprint('compat_event_registration', __name__, url_prefix='/event/<int:event_id>')
 _compat_bp.add_url_rule('/registration/', 'registration', compat_registration)
